Add rule status filter to the recs list table
