+++ cypress/fixtures/RecsListTable/data.json
@@ -9,7 +9,7 @@
       "impact": 1,
       "likelihood": 4,
       "tags": ["openshift", "etcd", "performance"],
-      "rule_status": "enabled",
+      "disabled": false,
       "risk_of_change": 0,
       "impacted_clusters_count": 1
     },
@@ -22,7 +22,7 @@
       "impact": 2,
       "likelihood": 2,
       "tags": ["openshift", "configuration", "service_availability"],
-      "rule_status": "enabled",
+      "disabled": false,
       "risk_of_change": 0,
       "impacted_clusters_count": 42
     },
@@ -35,7 +35,7 @@
       "impact": 2,
       "likelihood": 3,
       "tags": ["openshift", "configuration", "service_availability"],
-      "rule_status": "enabled",
+      "disabled": false,
       "risk_of_change": 0,
       "impacted_clusters_count": 0
     },
@@ -48,7 +48,7 @@
       "impact": 1,
       "likelihood": 4,
       "tags": ["service_availability", "vsphere", "disk"],
-      "rule_status": "disabled",
+      "disabled": false,
       "risk_of_change": 0,
       "impacted_clusters_count": 0
     },
@@ -61,7 +61,7 @@
       "impact": 2,
       "likelihood": 3,
       "tags": ["openshift", "performance", "osd_customer"],
-      "rule_status": "disabled",
+      "disabled": false,
       "risk_of_change": 0,
       "impacted_clusters_count": 0
     },
@@ -74,7 +74,7 @@
       "impact": 2,
       "likelihood": 2,
       "tags": ["service_availability", "openshift", "operator"],
-      "rule_status": "enabled",
+      "disabled": false,
       "risk_of_change": 0,
       "impacted_clusters_count": 100
     },
@@ -87,33 +87,33 @@
       "impact": 1,
       "likelihood": 2,
       "tags": ["openshift", "service_availability"],
-      "rule_status": "enabled",
+      "disabled": false,
       "risk_of_change": 0,
       "impacted_clusters_count": 2003
     },
     {
       "rule_id": "external.rules.ru123le_n_one|ERROR_321KEY_N1",
-      "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+      "description": "disabled rule with 2 impacted",
       "generic": "Sed ut perspiciatis unde omnis iste natus error. \n",
       "publish_date": "2021-02-17T15:00:00Z",
       "total_risk": 2,
       "impact": 1,
       "likelihood": 4,
       "tags": ["openshift", "service_availability"],
-      "rule_status": "enabled",
+      "disabled": true,
       "risk_of_change": 0,
       "impacted_clusters_count": 2
     },
     {
       "rule_id": "external.rules.rules.rule_n_eleventh|ERROR_KEY_11",
-      "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+      "description": "disabled rule with 0 impacted",
       "generic": "Sed ut perspiciatis unde omnis iste natus error. \n",
       "publish_date": "2021-02-17T15:00:00Z",
       "total_risk": 1,
       "impact": 1,
       "likelihood": 1,
       "tags": ["openshift", "service_availability"],
-      "rule_status": "enabled",
+      "disabled": true,
       "risk_of_change": 0,
       "impacted_clusters_count": 0
     }
